Unsigned integer instead of tiny integer for database compatability

+++ database/migrations/2019_05_21_092047_update_courses_table.php
@@ -15,7 +15,7 @@ class UpdateCoursesTable extends Migration
     {
         Schema::table('courses', function (Blueprint $table) {
             $table->string('description', 2048);
-            $table->unsignedTinyInteger('difficulty')->change();
+            $table->unsignedInteger('difficulty')->change();
         });
     }
